requirements.txt
accelerate==0.31.0
beautifulsoup4==4.12.3
chromadb==0.5.0
click==8.1.7
keras-nlp==0.12.1
pillow==10.3.0
praw==7.7.1
pyarrow==16.1.0
requests==2.31.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
tensorflow==2.16.1
torch==2.3.1
transformers==4.41.2
